requirements.txt
GitPython
markdown